% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{geotagged_photo_paths}
\alias{geotagged_photo_paths}
\title{Return file paths of small geotagged photos, shipped with package}
\usage{
geotagged_photo_paths()
}
\description{
Return file paths of small geotagged photos, shipped with package
}
