% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geotagged_photos}
\alias{geotagged_photos}
\title{Stats19 schema and variables}
\description{
\code{geotagged_photos}is a list of \code{magick-image} objects, representing
geotagged photos generated with the following command:
\code{geotagged_photos = make_icons(geotagged_photo_paths())}.
}
\details{
See the \code{data-raw} folder in the repo for details.
}
\keyword{datasets}
