% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{make_icons}
\alias{make_icons}
\title{Generate a list of icon images}
\usage{
make_icons(f, geometry = "400x300")
}
\arguments{
\item{f}{File path to photo(s).}

\item{geometry}{A geometry specification such as \code{"100x80"}, meaning 100 pixels wide, 80 pixels high.
See \code{\link[magick:geometry]{magick::geometry()}} for details.}
}
\description{
Generate a list of icon images
}
