% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_photos.R
\name{photo_add_captions}
\alias{photo_add_captions}
\title{Add photo captions with preview}
\usage{
photo_add_captions(f, geometry = "800x600")
}
\arguments{
\item{f}{File path to photo(s).}

\item{geometry}{A geometry specification such as \code{"100x80"}, meaning 100 pixels wide, 80 pixels high.
See \code{\link[magick:geometry]{magick::geometry()}} for details.}
}
\description{
Add photo captions with preview
}
\examples{
if(interactive()) {
photo_add_captions(geotagged_photo_paths())
}
}
