% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{photo_compress_width}
\alias{photo_compress_width}
\title{Compress photos}
\usage{
photo_compress_width(f, width = 800, quality = 70, ...)
}
\arguments{
\item{f}{File path to photo(s).}

\item{width}{The width of the results; aspect ratio will remain unchanged}

\item{quality}{Number from 0 to 100 representing the quality of .jpg images}

\item{...}{Other arguments to pass to \code{\link[magick:image_write]{magick::image_write()}}}
}
\description{
Warning: \strong{overwrites photos}
}
