% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photos.R
\name{photo_group}
\alias{photo_group}
\title{Group photos by time and location}
\usage{
photo_group(photo, n_groups = 3, date_time_cut = NULL,
  cluster_string = "Latitude$|Longitude$|date_time")
}
\arguments{
\item{photo}{An sf object created by \code{photos_sf}}

\item{n_groups}{The number of groups if \code{date_time_cut} is not provided}

\item{date_time_cut}{Times provided to break the photos}

\item{cluster_string}{Text string used to select variables used for clustering}
}
\description{
This function allows you to group photos, by time or a combination of time and location.
If \code{date_time_cut} values are taken (times in which to break the photos)
time alone will be used to break-up the photos.
If not, you can simply request a number of breaks with \code{n_groups}.
}
\examples{
f = geotagged_photo_paths()
photo = photos_sf(f)
photo_group(photo, 2)
summary(photo$date_time)
date_time_cut = lubridate::ymd_hm(c("2019-05-23 00:00", "2019-05-24 11:00", "2019-05-24 24:00"))
photo$group = photo_group(photo, date_time_cut = date_time_cut)
plot(photo["DateTimeOriginal"], pch = photo$group)
}
