% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{photo_iconize}
\alias{photo_iconize}
\title{Save icons of photos}
\usage{
photo_iconize(f, geometry = "400x300", dir_out = tempdir(),
  format = "png", overwrite = TRUE, ...)
}
\arguments{
\item{f}{File path to photo(s).}

\item{geometry}{A geometry specification such as \code{"100x80"}, meaning 100 pixels wide, 80 pixels high.
See \code{\link[magick:geometry]{magick::geometry()}} for details.}

\item{dir_out}{Directory where files should be saved, \code{tempdir()} by default}

\item{format}{Output format. See \code{\link[magick:image_write]{magick::image_write()}} for details}

\item{overwrite}{Should previous files be overwritten? \code{TRUE} by default}

\item{...}{Other arguments to pass to \code{\link[magick:image_write]{magick::image_write()}}}
}
\description{
Save icons of photos
}
