% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_photos.R
\name{photo_select}
\alias{photo_select}
\title{Helper function to select photos}
\usage{
photo_select(f, geometry = "800x600")
}
\arguments{
\item{f}{File path to photo(s).}

\item{geometry}{A geometry specification such as \code{"100x80"}, meaning 100 pixels wide, 80 pixels high.
See \code{\link[magick:geometry]{magick::geometry()}} for details.}
}
\description{
Helper function to select photos
}
\examples{
if(interactive()) {
select_photos(geotagged_photo_paths())
}
}
