% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photobook.R
\name{photobook}
\alias{photobook}
\title{Create photobook index}
\usage{
photobook(photo, index = photobook_index(), groups = NULL,
  n_groups = 3, captions = NULL, dir_out = tempdir())
}
\arguments{
\item{photo}{An sf object created by \code{photos_sf}}

\item{index}{A character string produced by \code{photobook_index}}

\item{groups}{Character vector the same length as \code{nrow(photo)}}

\item{n_groups}{The number of groups if \code{date_time_cut} is not provided}

\item{captions}{A character vector of captions}

\item{dir_out}{Directory where files should be saved, \code{tempdir()} by default}
}
\description{
Create photobook index
}
\examples{
f = geotagged_photo_paths()

photo = photos_sf(f)
photobook(photo)
\donttest{
# file.edit(file.path(tempdir(), "index.Rmd"))
setwd(tempdir())
list.files()
bookdown::render_book("index.Rmd")
browseURL("_book/index.html")
}
}
