% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photobook.R
\name{photobook_index}
\alias{photobook_index}
\title{Create an index.Rmd file}
\usage{
photobook_index(photobook_name = "Name of the book",
  photobook_author = "Names of the creators",
  photobook_repo = "org/repo", photobook_description = "Photobook.")
}
\arguments{
\item{photobook_name}{Name of the book}

\item{photobook_author}{Names of the creators}

\item{photobook_repo}{github repo, e.g. `"org/repo"``}

\item{photobook_description}{Photobook description.}
}
\description{
Create an index.Rmd file
}
