% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photomap.R
\name{photomap}
\alias{photomap}
\title{Creates an interactive photomap}
\usage{
photomap(f, icons = NULL, dir_out = tempdir(), iconWidth = 100,
  iconHeight = 75)
}
\arguments{
\item{f}{File path to photo(s).}

\item{icons}{File paths to icons of photos}

\item{dir_out}{Directory where files should be saved, \code{tempdir()} by default}

\item{iconWidth}{Width of the photo icon, in pixels}

\item{iconHeight}{Height of the photo icon, in pixels}
}
\description{
Creates an interactive photomap
}
\examples{
f = geotagged_photo_paths()
photomap(f)
}
