% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_photos.R
\name{select_photos}
\alias{select_photos}
\title{Make a slideshow of photos}
\usage{
select_photos(f, geometry = "800x600")
}
\arguments{
\item{f}{File path to photo(s).}

\item{geometry}{A geometry specification such as \code{"100x80"}, meaning 100 pixels wide, 80 pixels high.
See \code{\link[magick:geometry]{magick::geometry()}} for details.}
}
\description{
Make a slideshow of photos
}
\examples{
if(interactive()) {
select_photos(geotagged_photo_paths())
}
}
