% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideshow.R
\name{slideshow}
\alias{slideshow}
\title{Make a slideshow of photos}
\usage{
slideshow(f, delay = 1, geometry = "800x600", repeat_show = FALSE)
}
\arguments{
\item{f}{File path to photo(s).}

\item{delay}{Delay in seconds between photos to show}

\item{geometry}{A geometry specification such as \code{"100x80"}, meaning 100 pixels wide, 80 pixels high.
See \code{\link[magick:geometry]{magick::geometry()}} for details.}

\item{repeat_show}{If \code{TRUE} the photos will repeat in an infinite loop. \code{FALSE} by default.}
}
\description{
Make a slideshow of photos
}
\examples{
slideshow(geotagged_photo_paths())
}
